requirements.txt
# Core dependencies
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
pyyaml>=6.0

# OCR and image processing
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
opencv-python>=4.8.0
python-magic>=0.4.27

# Document processing
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI/ML
numpy>=1.24.0
scikit-learn>=1.3.0

# Async support
aiohttp>=3.8.0

# Development dependencies (install with: pip install -e ".[dev]")
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Bank Statement Processing (added 2025-11-06)
lxml>=4.9.0
defusedxml>=0.7.0
python-dateutil>=2.8.0

# PDF OCR Layer (added 2025-11-06)
PyMuPDF>=1.23.0  # fitz
